feat 자동차 이름 검증 기능 구현
